Refactor: Add useTryIt hook
